Separate domain and persistence logic into sep. funcs
